pyproject.toml
[project]
name = "internshipprojectdrf"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django>=5.2.7",
    "django-filter>=25.1",
    "django-jazzmin>=3.0.1",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.5.1",
    "drf-yasg>=1.21.11",
    "pillow>=11.3.0",
    "psycopg[binary]>=3.2.10",
    "python-dotenv>=1.1.1",
]
